pyproject.toml
[tool.poetry]
name = "web-dev"
version = "0.1.0"
description = "A very cool web app"
authors = ["Gabriel Briones <gabriel.briones@oracle.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "1.3.23"
aioredis = "^2.0.1"
Jinja2 = "^3.1.2"
alembic = "^1.7.7"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
